old_projects/bot_helper/requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.7.9
charset-normalizer==3.4.2
colorama==0.4.6
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
loguru==0.7.3
magic-filter==1.0.12
multidict==6.6.3
playwright==1.53.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
python-dotenv==1.1.1
requests==2.32.4
soupsieve==2.7
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
win32_setctime==1.2.0
yarl==1.20.1
